Add files via upload
In learning with data that is high dimensional, the grouping of characteristics is quite useful. The techniques described in this literature improve feature selection stability and reduces estimation variance, resulting in better model accuracy and also assisting with data interpretation. One such tool is discussed for selecting variables and arranging them into prediction clusters at the same time which shows good results. In addition to increasing the accuracy of predictions and interpretation. The purpose of this study is to compare ridge, lasso, elastic-net, group lasso, OSCAR (octagonal shrinkage and clustering algorithm for regression), and linear regression models to see which one works best as demonstrated by experimental results on simulated and soil data.